requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers
mysqlclient>=2.2.0
django-extensions
djangorestframework-simplejwt